Display tracklist in Info Dialog Infos tab for albums too
On not-yet-matched clusters, context menu shows Infos item and user can see a copiable tracklist in the Info tab.
On loaded albums the context menu will be there only when album has cover art (and no errors).

This patch unifies things a bit without adding much code, Infos dialog is accessible as soon an album is loaded, and Infos tab displays a text tracklist for it, which can be convenient for the user as it is copiable.